// demos/field-service-dispatch/src/generated/models/Fsd_statuslogsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Fsd_statuslogsfsd_status = {
  1: 'New',
  2: 'Assigned',
  3: 'EnRoute',
  4: 'OnSite',
  5: 'Completed',
  6: 'Cancelled'
} as const;
export type Fsd_statuslogsfsd_status = keyof typeof Fsd_statuslogsfsd_status;
export const Fsd_statuslogsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Fsd_statuslogsstatecode = keyof typeof Fsd_statuslogsstatecode;
export const Fsd_statuslogsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Fsd_statuslogsstatuscode = keyof typeof Fsd_statuslogsstatuscode;

export interface Fsd_statuslogsBase {
  fsd_loggedon?: string;
  fsd_name?: string;
  fsd_notes?: string;
  fsd_status?: Fsd_statuslogsfsd_status;
  fsd_statuslogid: string;
  fsd_technicianname?: string;
  "fsd_ticketid@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Fsd_statuslogsstatecode;
  statuscode?: Fsd_statuslogsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Fsd_statuslogs extends Fsd_statuslogsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  fsd_statusname?: string;
  fsd_ticketidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  fsd_ticketid?: object;
  _fsd_ticketid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}